requirements.txt
discord.py>=0.16.12
praw>=5.3.0
apscheduler>=3.5.0
psycopg2>=2.7.3.2